Add Ui, Storage, Parser, TaskList, Command
