+++ src/main/java/Command.java
@@ -1,0 +1,11 @@
+public enum Command {
+    BYE,
+    LIST,
+    DELETE,
+    TODO,
+    DEADLINE,
+    EVENT,
+    MARK,
+    UNMARK,
+    INVALID
+}
