+++ src/main/java/TaskList.java
@@ -1,24 +1,28 @@
 import java.util.ArrayList;
 
-public class Memory {
+public class TaskList {
     
     private ArrayList<Task> tasks;
     private int numTasks;
+    private boolean isSaved;
 
-    Memory() {
+    TaskList() {
         this.tasks = new ArrayList<>();
         this.numTasks = 0;
+        this.isSaved = false;
     }
 
-    public boolean add(Task task) {
+    public boolean addTask(Task task) {
         this.tasks.add(task);
         this.numTasks++;
+        this.isSaved = false;
         return true;
     }
 
     public Task deleteTask(int index) {
         Task task = this.tasks.remove(index);
         this.numTasks--;
+        this.isSaved = false;
         return task;
     }
 
@@ -33,17 +37,27 @@ public class Memory {
     public Task markTask(int index) {
         Task task = this.tasks.get(index);
         task.mark();
+        this.isSaved = false;
         return task;
     }
 
     public Task unmarkTask(int index) {
         Task task = this.tasks.get(index);
         task.unmark();
+        this.isSaved = false;
         return task;
     }
 
     public int getNumTasks() {
         return this.numTasks;
+    }
+
+    public boolean isSaved() {
+        return this.isSaved;
+    }
+
+    public void save() {
+        this.isSaved = true;
     }
 
 }
